requirements.txt

numpy==1.26.4
opencv-python==4.10.0.84
imutils==0.5.4
pyyaml==6.0.2
pyttsx3==2.90
twilio==9.3.3
